+++ 1.21.8/Common/src/main/java/fuzs/pickupnotifier/client/gui/entry/DisplayEntry.java
@@ -16,23 +16,30 @@ import net.minecraft.world.item.Rarity;
 import org.jetbrains.annotations.MustBeInvokedByOverriders;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-public abstract class DisplayEntry {
-    public static final int ENTRY_HEIGHT = 18;
+public abstract class DisplayEntry<T> {
+    public static final int ELEMENT_HEIGHT = 18;
     private static final int TEXT_ITEM_MARGIN = 4;
 
-    private final Rarity rarity;
-    private int remainingTicks;
-    private int displayAmount;
-    private Component component;
+    protected final T item;
+    private final int displayAmount;
+    private final Component component;
+    private int remainingTicks = PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.displayTime;
 
-    public DisplayEntry(int displayAmount, Rarity rarity) {
+    public DisplayEntry(T item, int displayAmount, Rarity rarity) {
+        this.item = item;
         this.displayAmount = displayAmount;
-        this.rarity = rarity;
-        this.resetEntry();
+        this.component = Component.empty().append(this.createTextComponent()).setStyle(this.getComponentStyle(rarity));
     }
+
+    public Object getKey() {
+        return this;
+    }
+
+    protected abstract Component getEntryName(T item);
 
     public int getDisplayAmount() {
         return this.displayAmount;
@@ -42,34 +49,18 @@ public abstract class DisplayEntry {
         return this.remainingTicks <= 0;
     }
 
+    @MustBeInvokedByOverriders
     public void tick() {
         if (this.remainingTicks > 0) {
             this.remainingTicks--;
         }
     }
 
-    protected abstract Component getEntryName();
-
-    public Component getTextComponent() {
-
-        if (this.component == null) {
-
-            this.component = Component.empty()
-                    .append(this.createTextComponent(PickUpNotifier.CONFIG.get(ClientConfig.class).display.position.mirrored()))
-                    .setStyle(this.getComponentStyle());
-        }
-
-        return this.component;
-    }
-
-    private Component createTextComponent(boolean reverse) {
-
+    protected final Component createTextComponent() {
+        boolean reverse = PickUpNotifier.CONFIG.get(ClientConfig.class).display.position.isRight();
         List<Component> components = new ArrayList<>();
-
         this.appendTextComponents(components, reverse);
-
         if (reverse) {
-
             Collections.reverse(components);
         }
 
@@ -84,63 +75,49 @@ public abstract class DisplayEntry {
     @MustBeInvokedByOverriders
     protected void appendTextComponents(List<Component> components, boolean reverse) {
         if (PickUpNotifier.CONFIG.get(ClientConfig.class).display.displayItemName) {
-
-            components.add(this.getEntryName());
+            components.add(this.getEntryName(this.item));
         }
 
-        int displayAmount =
-                PickUpNotifier.CONFIG.get(ClientConfig.class).display.displayAmount.isText() ? this.getDisplayAmount() :
-                        0;
-        if (displayAmount > 1
-                || displayAmount == 1 && PickUpNotifier.CONFIG.get(ClientConfig.class).display.displaySingleCount) {
-
-            components.add(Component.literal(reverse ? displayAmount + "x" : "x" + displayAmount));
+        if (PickUpNotifier.CONFIG.get(ClientConfig.class).display.displayAmount.isText()) {
+            if (this.displayAmount > 1 || this.displayAmount == 1
+                    && PickUpNotifier.CONFIG.get(ClientConfig.class).display.displaySingleCount) {
+                components.add(Component.literal(reverse ? this.displayAmount + "x" : "x" + this.displayAmount));
+            }
         }
     }
 
-    private Style getComponentStyle() {
-
-        if (!PickUpNotifier.CONFIG.get(ClientConfig.class).display.ignoreRarity && this.rarity != Rarity.COMMON) {
-
-            return ItemHelper.getRarityStyle(this.rarity);
+    private Style getComponentStyle(Rarity rarity) {
+        if (!PickUpNotifier.CONFIG.get(ClientConfig.class).display.ignoreRarity && rarity != Rarity.COMMON) {
+            return ItemHelper.getRarityStyle(rarity);
         } else {
-
             return Style.EMPTY.withColor(PickUpNotifier.CONFIG.get(ClientConfig.class).display.textColor);
         }
     }
 
-    public float getRemainingTicksRelative(float partialTicks) {
-
-        float moveTime = Math.min(PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.moveTime,
-                PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.displayTime);
-        return 1.0F - Mth.clamp((this.remainingTicks - partialTicks) / moveTime, 0.0F, 1.0F);
+    public static double getRelativeRemainingTicks(Collection<DisplayEntry<?>> values, float partialTick) {
+        return values.stream()
+                .mapToDouble((DisplayEntry<?> displayEntry) -> 1.0
+                        - displayEntry.getRelativeRemainingTicks(partialTick))
+                .average()
+                .orElse(0.0) * values.size();
     }
 
-    public void resetEntry() {
-
-        this.remainingTicks = PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.displayTime;
-        this.component = null;
+    public final float getRelativeRemainingTicks(float partialTick) {
+        float moveTime = PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.getMoveTime();
+        return Mth.clamp((this.remainingTicks - partialTick) / moveTime, 0.0F, 1.0F);
     }
 
-    public abstract boolean mayMergeWith(DisplayEntry other, boolean excludeNamed);
-
-    public void mergeWith(DisplayEntry other) {
-
-        this.displayAmount = this.displayAmount + other.displayAmount;
-        this.resetEntry();
-    }
+    public abstract DisplayEntry<?> mergeWith(DisplayEntry<?> otherDisplayEntry);
 
     public int getEntryWidth(Font font) {
-
-        int textWidth = font.width(this.getTextComponent());
+        int textWidth = font.width(this.component);
         return PickUpNotifier.CONFIG.get(ClientConfig.class).display.drawSprite ?
                 textWidth + (textWidth == 0 ? 0 : TEXT_ITEM_MARGIN) + 16 : textWidth;
     }
 
     public void render(GuiGraphics guiGraphics, Font font, int posX, int posY, float alpha) {
-
-        float scale = PickUpNotifier.CONFIG.get(ClientConfig.class).display.getScale();
-        boolean mirrorPosition = PickUpNotifier.CONFIG.get(ClientConfig.class).display.position.mirrored();
+        float scale = PickUpNotifier.CONFIG.get(ClientConfig.class).display.getDisplayScale();
+        boolean mirrorPosition = PickUpNotifier.CONFIG.get(ClientConfig.class).display.position.isRight();
         boolean withSprite = PickUpNotifier.CONFIG.get(ClientConfig.class).display.drawSprite;
         int textStartX = mirrorPosition || !withSprite ? posX : posX + 16 + TEXT_ITEM_MARGIN;
 
@@ -148,30 +125,24 @@ public abstract class DisplayEntry {
         guiGraphics.pose().scale(scale, scale);
 
         this.renderBg(guiGraphics, font, posX, posY, alpha);
-
-        float fadeTime = PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.fadeAway ? 1.0F - alpha : 1.0F;
-        guiGraphics.drawString(font, this.getTextComponent(), textStartX, posY + 4, ARGB.white(fadeTime), true);
+        guiGraphics.drawString(font, this.component, textStartX, posY + 4, ARGB.white(alpha), true);
         if (withSprite) {
-
-            int textWidth = font.width(this.getTextComponent());
+            int textWidth = font.width(this.component);
             this.renderSprite(guiGraphics,
                     font,
                     mirrorPosition ? posX + textWidth + (textWidth == 0 ? 0 : TEXT_ITEM_MARGIN) : posX,
                     posY,
-                    fadeTime);
+                    alpha);
         }
 
         guiGraphics.pose().popMatrix();
     }
 
     private void renderBg(GuiGraphics guiGraphics, Font font, int posX, int posY, float alpha) {
-
         switch (PickUpNotifier.CONFIG.get(ClientConfig.class).display.entryBackground) {
-
             case CHAT -> {
-
                 int backgroundOpacity = ARGB.color(ARGB.as8BitChannel(Mth.clamp(
-                        Minecraft.getInstance().options.textBackgroundOpacity().get().floatValue() * (1.0F - alpha),
+                        Minecraft.getInstance().options.textBackgroundOpacity().get().floatValue() * alpha,
                         0.0F,
                         1.0F)), 0);
                 int endY = posY + 16;
@@ -181,18 +152,22 @@ public abstract class DisplayEntry {
 
                 guiGraphics.fill(posX - 3, posY, posX + this.getEntryWidth(font) + 5, endY, backgroundOpacity);
             }
-
             case TOOLTIP -> {
-
                 DisplayEntryRenderHelper.renderTooltipBackground(guiGraphics,
                         posX,
                         posY + 3,
                         this.getEntryWidth(font),
                         9,
-                        ARGB.white(1.0F - alpha));
+                        ARGB.white(alpha));
             }
         }
     }
 
-    protected abstract void renderSprite(GuiGraphics guiGraphics, Font font, int posX, int posY, float fadeTime);
+    protected abstract void renderSprite(GuiGraphics guiGraphics, Font font, int posX, int posY, float alpha);
+
+    @Override
+    public abstract boolean equals(Object obj);
+
+    @Override
+    public abstract int hashCode();
 }