+++ 1.21.8/Common/src/main/java/fuzs/pickupnotifier/client/gui/entry/ItemDisplayEntry.java
@@ -3,108 +3,72 @@ package fuzs.pickupnotifier.client.gui.entry;
 import fuzs.pickupnotifier.PickUpNotifier;
 import fuzs.pickupnotifier.client.util.DisplayEntryRenderHelper;
 import fuzs.pickupnotifier.config.ClientConfig;
+import fuzs.pickupnotifier.config.CombineEntries;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.ContainerHelper;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.item.ItemStack;
 
 import java.util.List;
+import java.util.Objects;
 
-public class ItemDisplayEntry extends DisplayEntry {
-    static final int ITEM_STACK_POP_TIME = 5;
+public final class ItemDisplayEntry extends DisplayEntry<ItemStack> {
+    private static final int ITEM_STACK_POP_TIME = 5;
 
-    private final ItemStack itemStack;
-
-    public ItemDisplayEntry(ItemStack itemStack, int amount) {
-        super(amount, itemStack.getRarity());
-        this.itemStack = createDisplayItemStack(itemStack);
+    public ItemDisplayEntry(ItemStack itemStack, int displayAmount) {
+        super(itemStack.copy(), displayAmount, itemStack.getRarity());
+        this.item.setPopTime(ITEM_STACK_POP_TIME);
     }
 
-    private static ItemStack createDisplayItemStack(ItemStack itemStack) {
-        itemStack = itemStack.copy();
-        // remove enchantments from copy as we don't want the glint to show
-        if (PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.combineEntries
-                == ClientConfig.CombineEntries.ALWAYS) {
-            itemStack.set(DataComponents.ENCHANTMENT_GLINT_OVERRIDE, false);
+    @Override
+    protected Component getEntryName(ItemStack itemStack) {
+        if (PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.combineEntries == CombineEntries.ALWAYS) {
+            return itemStack.getItem().getName(itemStack);
+        } else {
+            return itemStack.getHoverName();
         }
-
-        itemStack.setPopTime(ITEM_STACK_POP_TIME);
-        return itemStack;
     }
 
     @Override
     public void tick() {
         super.tick();
-        if (this.itemStack.getPopTime() > 0) {
-            this.itemStack.setPopTime(this.itemStack.getPopTime() - 1);
+        if (this.item.getPopTime() > 0) {
+            this.item.setPopTime(this.item.getPopTime() - 1);
         }
     }
 
     @Override
-    protected Component getEntryName() {
-        if (PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.combineEntries
-                == ClientConfig.CombineEntries.ALWAYS) {
-            return this.itemStack.getItem().getName(this.itemStack);
-        } else {
-            return this.itemStack.getHoverName();
-        }
-    }
-
-    @Override
-    public boolean mayMergeWith(DisplayEntry other, boolean excludeNamed) {
-        return other instanceof ItemDisplayEntry itemDisplayEntry && isSameItemSameName(this.itemStack,
-                itemDisplayEntry.itemStack,
-                excludeNamed);
-    }
-
-    @Override
-    public void mergeWith(DisplayEntry other) {
-        super.mergeWith(other);
-        this.itemStack.setPopTime(ITEM_STACK_POP_TIME);
-    }
-
-    static boolean isSameItemSameName(ItemStack itemStack, ItemStack otherItemStack, boolean preferHoverName) {
-        // TODO replace this with hashing and a linked hash map in favor of a normal list
-        if (ItemStack.isSameItem(itemStack, otherItemStack)) {
-            if (preferHoverName) {
-                return itemStack.getHoverName().equals(otherItemStack.getHoverName()) && itemStack.getRarity()
-                        .equals(otherItemStack.getRarity());
-            } else {
-                return itemStack.getItem().getName(itemStack).equals(otherItemStack.getItem().getName(otherItemStack));
-            }
-        } else {
-            return false;
-        }
+    public DisplayEntry<?> mergeWith(DisplayEntry<?> otherDisplayEntry) {
+        return new ItemDisplayEntry(this.item, this.getDisplayAmount() + otherDisplayEntry.getDisplayAmount());
     }
 
     @Override
     protected void appendTextComponents(List<Component> components, boolean reverse) {
         if (PickUpNotifier.CONFIG.get(ClientConfig.class).display.inventoryCount) {
-            int inventoryCount = this.getInventoryCount(Minecraft.getInstance().player.getInventory());
-            Component component = Component.literal(String.valueOf(inventoryCount));
-            components.add(Component.literal("(").append(component).append(")"));
+            Component inventoryCountComponent = Component.literal(String.valueOf(this.getInventoryCount()));
+            components.add(Component.literal("(").append(inventoryCountComponent).append(")"));
         }
 
         super.appendTextComponents(components, reverse);
     }
 
-    private int getInventoryCount(Inventory inventory) {
+    private int getInventoryCount() {
+        Inventory inventory = Minecraft.getInstance().player.getInventory();
         return ContainerHelper.clearOrCountMatchingItems(inventory,
-                (ItemStack itemStack) -> ItemStack.isSameItem(this.itemStack, itemStack),
+                (ItemStack itemStackX) -> ItemStack.isSameItem(this.item, itemStackX),
                 0,
                 true);
     }
 
     @Override
-    protected void renderSprite(GuiGraphics guiGraphics, Font font, int posX, int posY, float fadeTime) {
-        if (this.itemStack.getPopTime() > 0) {
+    protected void renderSprite(GuiGraphics guiGraphics, Font font, int posX, int posY, float alpha) {
+        if (this.item.getPopTime() > 0) {
             DeltaTracker deltaTracker = Minecraft.getInstance().getDeltaTracker();
-            float popTime = this.itemStack.getPopTime() - deltaTracker.getGameTimeDeltaPartialTick(false);
+            float popTime = this.item.getPopTime() - deltaTracker.getGameTimeDeltaPartialTick(false);
             float popTimeScale = 1.0F + popTime / ITEM_STACK_POP_TIME;
             guiGraphics.pose().pushMatrix();
             guiGraphics.pose().translate(posX + 8, posY + 12);
@@ -112,8 +76,8 @@ public class ItemDisplayEntry extends DisplayEntry {
             guiGraphics.pose().translate(-(posX + 8), -(posY + 12));
         }
 
-        DisplayEntryRenderHelper.renderItem(guiGraphics, this.itemStack, posX, posY, fadeTime);
-        if (this.itemStack.getPopTime() > 0) {
+        DisplayEntryRenderHelper.renderItem(guiGraphics, this.item, posX, posY, alpha);
+        if (this.item.getPopTime() > 0) {
             guiGraphics.pose().popMatrix();
         }
 
@@ -123,7 +87,34 @@ public class ItemDisplayEntry extends DisplayEntry {
                     this.getDisplayAmount(),
                     posX,
                     posY,
-                    fadeTime);
+                    alpha);
         }
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == this) {
+            return true;
+        } else if (PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.combineEntries != CombineEntries.NEVER
+                && obj instanceof ItemDisplayEntry itemDisplayEntry) {
+            if (ItemStack.isSameItem(this.item, itemDisplayEntry.item) && Objects.equals(this.item.getRarity(),
+                    itemDisplayEntry.item.getRarity()) && Objects.equals(this.item.isEnchanted(),
+                    itemDisplayEntry.item.isEnchanted())) {
+                if (PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.combineEntries == CombineEntries.ALWAYS) {
+                    return true;
+                } else {
+                    return Objects.equals(this.item.getHoverName(), itemDisplayEntry.item.getHoverName());
+                }
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
+    }
+
+    @Override
+    public int hashCode() {
+        return this.item.getItem().hashCode();
     }
 }