+++ 1.21.8/CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v21.8.1-1.21.8] - 2025-09-12
 
+### Added
+
+- Add five additional position presets (`TOP_CENTER`, `CENTER_LEFT`, `CENTER`, `CENTER_RIGHT`, `BOTTOM_CENTER`)
+- Add an option for entries to move out to the sides in addition to bottom / top
+- Add a pop animation when collecting new items
+
 ### Changed
 
 - Support fading for items once more
-- Add a pop animation when collecting new items
+- Improve support for enchanted items and the enchantment glint
 
 ## [v21.8.0-1.21.8] - 2025-07-18